Allow passing wasmHash as string for getting wasm bytecode
